--- ansible_galaxy/models/content_archive.py
@@ -1,15 +1,0 @@
-import attr
-
-
-@attr.s(frozen=True)
-class ContentArchiveMeta(object):
-    requires_meta_main = False
-
-    archive_type = attr.ib()
-    top_dir = attr.ib()
-    archive_path = attr.ib(default=None)
-
-    # download url?
-    # file path?
-    # checksum?
-    # signature?
