+++ ansible_galaxy/models/models.py
@@ -67,6 +67,20 @@ class CollectionArtifactManifest(object):
 
 
 @attr.s(frozen=True)
+class ContentArchiveMeta(object):
+    requires_meta_main = False
+
+    archive_type = attr.ib()
+    top_dir = attr.ib()
+    archive_path = attr.ib(default=None)
+
+    # download url?
+    # file path?
+    # checksum?
+    # signature?
+
+
+@attr.s(frozen=True)
 class GalaxyContentMeta(object):
     namespace = attr.ib()
     name = attr.ib()
